requirements.txt
pyspark
apache-airflow
snowflake-connector-python
pandas
python-dotenv
